+++ .gitignore
@@ -1,3 +1,4 @@
 dist
 dist/
 output.zip
+.env.rc
