+++ src/value_env_args.go
@@ -1,0 +1,13 @@
+package main
+
+import "os"
+
+type EnvArgs struct {
+	APIKey string
+}
+
+func NewEnvArgs() EnvArgs {
+	return EnvArgs{
+		APIKey: os.Getenv("DODO_API_KEY"),
+	}
+}
